refactor: removes repeated code from pages and turns into ChatLayout
