+++ packages/ui/app/layout.tsx
@@ -1,8 +1,9 @@
 import { Providers } from "./providers";
 
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Chatbot - Mapa do Acolhimento",
+  description:
+    "Assistente virtual (IAna) criada pelo Mapa do Acolhimento para auxiliar na capacitação das voluntárias",
 };
 
 export default function RootLayout({
