+++ packages/ui/app/page.tsx
@@ -1,31 +1,10 @@
 "use client";
 
 import { useState } from "react";
-import { HStack, VStack, Heading, Center, useConst } from "@chakra-ui/react";
-import { Chat, ChatIcon } from "@/components";
+import { useConst } from "@chakra-ui/react";
+import { Chat, ChatLayout } from "@/components";
 import { ChatGPTMessage } from "@/components/ChatLine";
 import { useSearchParams } from "next/navigation";
-
-function ChatHeader() {
-  return (
-    <HStack
-      w={"full"}
-      minH={16}
-      borderBottom="1px solid"
-      borderColor="brand.lightGray"
-      gap={4}
-      align={"center"}
-      pl={8}
-    >
-      <Center boxSize={7} borderRadius={"base"} bg="#FFDBDB">
-        <ChatIcon boxSize={3.5} aria-label="Icon de mensagem" />
-      </Center>
-      <Heading as={"h2"} color="brand.primary" size={"sm"} fontWeight="700">
-        Fale com a IAna
-      </Heading>
-    </HStack>
-  );
-}
 
 const initialMessages = (name: string): ChatGPTMessage[] => [
   {
@@ -95,28 +74,16 @@ function Home() {
   };
 
   return (
-    <VStack spacing="10em" justify={"center"}>
-      <VStack
-        w={{ base: "full", lg: "80%" }}
-        maxH={"3xl"}
-        bg="white"
-        border="1px solid #C4C4C4"
-        borderRadius="0.625rem"
-        boxShadow={"0px 4px 4px 0px rgba(0, 0, 0, 0.10)"}
-        pb={9}
-        spacing={0}
-      >
-        <ChatHeader />
-        <Chat
-          loading={loading}
-          sendMessage={sendMessage}
-          messages={messages}
-          setMessages={setMessages}
-          city={city}
-          showSuggestions
-        />
-      </VStack>
-    </VStack>
+    <ChatLayout>
+      <Chat
+        loading={loading}
+        sendMessage={sendMessage}
+        messages={messages}
+        setMessages={setMessages}
+        city={city}
+        showSuggestions
+      />
+    </ChatLayout>
   );
 }
