+++ packages/ui/components/index.ts
@@ -3,3 +3,4 @@ export { default as ChatLine } from "./ChatLine";
 export { default as ChatIcon } from "./ChatIcon";
 export { default as InputMessage } from "./InputMessage";
 export { default as ChatSuggestions } from "./ChatSuggestions";
+export { default as ChatLayout } from "./ChatLayout";
